requirements.txt
langchain
langchain-google-genai
pillow
python-dotenv
pypdf
faiss-cpu
